Updated Feeding Method
cupsToFeed now set to 7, and motorTime set to 20 (so the feeder drops
3/4ths of a bowl).

cupsToFeed should be changed to bowlsToFeed to be more clear in its
function.

motorTime should be a calculated variable, and a new variable --
bowlSize --  should be introduced, which is a measure in cups.

To make this script easier to modify, we should move all editable
variables to a separate configuration file. This config can also be a
python script that'll be called from the main process to set variables.

`if MOTORON:` now includes a `for` loop to run the motor with a pause
in-between bowls. At this current moment, the delay is set to 3 seconds,
but will be user-defineable in the future.